Split out createresume into seperate protocol
